[SEARCH-1430] Solr responde comparator refactor
